Add compiled/schemas/.keep to ensure required directory exists for `npm run build`
A simple clone of the repo fails during `npm run build` because the
`compiled/schemas` directory doesn't exist when `flatten-schemas.sh` tries
to make the version directory below that.  This commit adds a .keep
to the `compile/schema` directory in line with what is done to retain
other directories.

The alternative would be to update `flatten-schemas.sh` line 12 to use
`-p` to also make parent directories as needed.

## Test Plan

Run `npm run build` after creating `compiled/schemas` and check it now completes.